move colormode into components
