+++ site/src/ui/components/index.ts
@@ -8,3 +8,4 @@ export { Icon } from './Icon-component'
 export { BadgeLink, IconBadge } from './Badge'
 export { Dropdown } from './Dropdown'
 export { Button } from './Button'
+export { ColorMode } from './ColorMode'
