+++ site/src/ui/theme-context.tsx
@@ -5,8 +5,6 @@ import {
   getThemePreference,
   setThemePreference,
 } from './dark-mode'
-import { Button, Dropdown } from './components'
-import { Icon, IconName } from './components/Icon-component'
 
 export type Theme = 'light' | 'dark'
 
@@ -79,40 +77,5 @@ export const ThemeProvider = ({ children }: { children: React.ReactNode }) => {
     <ThemeContext.Provider value={{ theme, preference, setTheme }}>
       {children}
     </ThemeContext.Provider>
-  )
-}
-
-export const ColorMode = ({ className }: { className?: string }) => {
-  const { theme, preference, setTheme } = useTheme()
-
-  const options = [
-    { label: 'Light', value: 'light', icon: 'sun' },
-    { label: 'Dark', value: 'dark', icon: 'moon' },
-    { label: 'System', value: undefined, icon: 'gear' },
-  ] as const
-
-  if (!theme) {
-    return null
-  }
-
-  const icon: IconName = theme === 'light' ? 'sun' : 'moon'
-
-  return (
-    <Button
-      onClick={() => setTheme(theme === 'light' ? 'dark' : 'light')}
-      className={className}
-    >
-      <Icon name={icon} width={16} height={16} />
-    </Button>
-  )
-
-  return (
-    <Dropdown
-      ariaLabel="Select site theme"
-      selected={preference}
-      options={options}
-      onSelect={setTheme}
-      className={className}
-    />
   )
 }
